[SPOT-304][REFACTOR] 2차 기획으로 인한 불필요한 댓글 좋아요 및 싫어요 관련 로직 삭제
