--- src/main/java/com/example/spot/post/application/query/GetLikedPostCommentUseCase.java
@@ -1,10 +1,0 @@
-package com.example.spot.post.application.query;
-
-public interface GetLikedPostCommentUseCase {
-    long countByPostCommentIdAndIsLikedTrue(Long postCommentId);
-
-    boolean existsByMemberIdAndPostCommentIdAndIsLikedTrue(Long postCommentId);
-
-    boolean existsByMemberIdAndPostCommentIdAndIsLikedFalse(Long postCommentId);
-
-}
