+++ src/main/java/com/example/spot/post/infrastructure/jpa/schedule/PostScheduleRealTimeRepository.java
@@ -1,4 +1,4 @@
-package com.example.spot.post.infrastructure.jpa;
+package com.example.spot.post.infrastructure.jpa.schedule;
 
 import com.example.spot.post.domain.schedule.PostScheduleRealTime;
 import org.springframework.data.jpa.repository.JpaRepository;
