+++ src/test/java/com/example/spot/service/post/PostCommandServiceTest.java
@@ -13,7 +13,6 @@ import com.example.spot.post.application.command.impl.LikePostUseCaseImpl;
 import com.example.spot.post.application.command.impl.ManagePostCommentUseCaseImpl;
 import com.example.spot.post.application.command.impl.ManagePostUseCaseImpl;
 import com.example.spot.post.application.command.impl.ScrapPostUseCaseImpl;
-import com.example.spot.post.application.query.GetLikedPostCommentUseCase;
 import com.example.spot.post.application.query.GetLikedPostUseCase;
 import com.example.spot.post.domain.Post;
 import com.example.spot.post.domain.PostComment;
@@ -21,7 +20,6 @@ import com.example.spot.post.domain.association.LikedPost;
 import com.example.spot.post.domain.association.LikedPostComment;
 import com.example.spot.post.domain.association.MemberScrap;
 import com.example.spot.post.domain.enums.Board;
-import com.example.spot.post.infrastructure.jpa.LikedPostCommentRepository;
 import com.example.spot.post.infrastructure.jpa.LikedPostRepository;
 import com.example.spot.post.infrastructure.jpa.MemberScrapRepository;
 import com.example.spot.post.infrastructure.jpa.PostCommentRepository;
@@ -77,16 +75,10 @@ class PostCommandServiceTest {
     private LikedPostRepository likedPostRepository;
 
     @Mock
-    private LikedPostCommentRepository likedPostCommentRepository;
-
-    @Mock
     private PostReportRepository postReportRepository;
 
     @Mock
     private GetLikedPostUseCase getLikedPostUseCase;
-
-    @Mock
-    private GetLikedPostCommentUseCase getLikedPostCommentUseCase;
 
     @InjectMocks
     private LikePostUseCaseImpl likePostUseCase;
@@ -149,24 +141,6 @@ class PostCommandServiceTest {
         when(likedPostRepository.existsByMemberIdAndPostId(2L, 2L)).thenReturn(false);
         when(getLikedPostUseCase.countByPostId(1L)).thenReturn(1L);
         when(getLikedPostUseCase.countByPostId(2L)).thenReturn(1L);
-
-        // LikedPostComment
-        when(likedPostCommentRepository.findByMemberIdAndPostCommentIdAndIsLikedFalse(1L, 1L))
-                .thenReturn(Optional.empty());
-        when(likedPostCommentRepository.findByMemberIdAndPostCommentIdAndIsLikedFalse(2L, 1L))
-                .thenReturn(Optional.of(member2LikedComment1));
-        when(likedPostCommentRepository.findByMemberIdAndPostCommentIdAndIsLikedTrue(1L, 1L))
-                .thenReturn(Optional.of(member1LikedComment1));
-        when(likedPostCommentRepository.findByMemberIdAndPostCommentIdAndIsLikedTrue(2L, 1L))
-                .thenReturn(Optional.empty());
-        when(likedPostCommentRepository.countByPostCommentIdAndIsLikedTrue(1L)).thenReturn(1L);
-        when(likedPostCommentRepository.countByPostCommentIdAndIsLikedTrue(2L)).thenReturn(0L);
-        when(likedPostCommentRepository.countByPostCommentIdAndIsLikedFalse(1L)).thenReturn(1L);
-        when(likedPostCommentRepository.countByPostCommentIdAndIsLikedFalse(2L)).thenReturn(0L);
-        when(likedPostCommentRepository.existsByMemberIdAndPostCommentIdAndIsLikedTrue(1L, 1L)).thenReturn(true);
-        when(likedPostCommentRepository.existsByMemberIdAndPostCommentIdAndIsLikedTrue(2L, 1L)).thenReturn(false);
-        when(likedPostCommentRepository.existsByMemberIdAndPostCommentIdAndIsLikedFalse(1L, 1L)).thenReturn(false);
-        when(likedPostCommentRepository.existsByMemberIdAndPostCommentIdAndIsLikedFalse(2L, 1L)).thenReturn(true);
 
         // MemberScrap
         when(memberScrapRepository.countByPostId(1L)).thenReturn(1L);
